Whee: style transitioning for focused nodes in AbstractVisualLayer.
To reduce the combinatorial explosion, there's no focused & pressed
state, instead a pressed state takes a precedence over focused. Which is
consistent with major UI toolkit styles I could check, and I feel that
while it is a limitation, it makes life easier for both the application
devs and the designers. And if such combined state is *really* desired,
there's always a possibility to implement a fully custom layer that
handles this -- the event classes expose this distinction.

As this is a breaking change for the setStyleTransition() API, I took
that as an opportunity to reorder the arguments in a way that makes more
sense. So watch out -- in addition to the function taking more
arguments, it now takes them in different order as well.