Whee: decouple parent node and top-level node order storage.
No functional change yet, i.e. it's still only possible to order root
nodes, it's just that the two things are no longer stored in the same
field.

The stress test gets ~5% slower due to the parent and order info now
being in two separate fields instead of one. The upcoming change may
change that either way, so not going to worry about that right now.