Whee: don't recycle node order slots in clearNodeOrder().
One vague reason would be that they'd be recycled right when putting
the hierarchies back to the order, so this avoid extra work. A
counterargument would be that in some cases the set of actually visible
top-level hierarchies will be significantly smaller than the set of
actually made hierarchies (think various dialogs, popups and whatnot, of
which only one is visible at a time), so the original approach would be
better in terms of memory consumption.

But the main reason is that with the introduction of non-root top-level
hierarchies, when they get disconnected from the top-level node
hierarchy, they should still remember how their children were ordered
relative to them, so e.g. when a dialog with a bunch of floating
child dropdowns and such is shown again, the application code doesn't
have to deal with the order on its own. And that is only achievable if
we don't recycle the order entry but just disconnect it (including the
children) from the outside.