Whee: assert that the handle is not null in fooHandle*Id().
This was originally not done because I abused NodeHandle to store both a
parent handle and the top-level node order slot index. But that's not
done anymore and because the *main* practical use of fooHandleId() is
for indexing into arrays, having it silently return 0 for null handles
is extremely, EXTREMELY error-prone, leading to either silent bugs, or
causing the caller code to contain an excessive number of assertions
validating that the handle is indeed not null.