Whee: fix BaseLayer edge artifacts.
In particular, the edges being cut off and the outline color leaking in
with non-zero smoothness values.

To avoid the cutoff, the quad is now expanded by the smoothness radius,
which subsequently means the gradient endpoints, texture coordinates and
also blur quads need to be adjusted as well. Because the texture
coordinates can be in a completely different scale than the actual
pixels, this is done on the CPU side in doUpdate() to not have to pass
extra scaling information to the shader.

To prevent the outline color leaks, the calculation is now changed so
that the outline color doesn't contribute at all if the outline width is
0, regardless of what's the difference between inner and outer
smoothness. This slightly expands the amount of work done in the
fragment shader, but I don't think it's too significant to be actually a
problem, and if it becomes there can always be an option to ignore the
inner smoothness and use the outer smoothness for both (which then
would make the code ~equivalent to before, just with a correct
calculation).

Test images are updated accordingly, the only differences are at the
edges, the rest is pixel-perfect match to the state before.