Whee: fully initialize the dynamic style in TextLayerStyleAnimator.
Before it was leaving the font, alignment and feature list in whatever
state it was before, causing strange inconsistency or even assertions
when a text with an animated style is set or edited. Such as when a
cursor is blinking.

Didn't realize this could be an issue until actually implementing a
blinking cursor and seeing how it behaves.