Whee: make smoothness pixel-density-dependent.
Instead of being in UI units. This means that the UI is perceivably
sharp the same way on both low-DPI and high-DPI systems without
it looking either blurry on high-DPI or aliased on low-DPI.

A consequence is that pixel density change now has to trigger update of
all the data (in BaseLayer w/o SubdividedQuads, at least), but that's a
relatively rare scenario to not be a problem -- the app isn't likely be
dragged between displays of different DPI too often.