Whee: setup renderer framebuffer directly in ui.setSize() always.
Originally I thought deferring to updateRenderer() would allow the
application to schedule heavy operations to a more appropriate time,
instead of suffering them directly when setSize() is called.

That however didn't really hold up in practice -- with a resizable
window and compositing / background blur enabled, this meant the
application couldn't just call renderer.compositingFramebuffer().clear()
at the beginning of draw(), it has to explicitly call updateRenderer()
before to avoid clearing a potentially stale framebuffer. In general,
any access to renderer.compositingFramebuffer() could possibly be stale,
hurting the UX a lot.

So, not doing that anymore. Everything is now done directly in
setSize().