Whee: require lay[out]er.setSize() to be called before update().
This was already required by the AbstractUserInterface::update() itself,
but as many tests call update() directly on layers / layouters, this
check was sidestepped. Then BaseLayer added its own assert (which was
impossible to get triggered by regular user code because the
AbstractUserInterface one would fire first), the assert didn't really
work if the blur scale vector was really tiny which was an annoying bug,
and doing it this way would mean each layer implementation would
eventually need to add the same check. Which doesn't scale.

So it's now done by the AbstractLayer / AbstractLayouter base directly,
which *may* be annoying for some tests, but definitely is cleaner and less
accident-prone.

Also, it's uncanny how close AbstractLayouter internals are close to
AbstractLayer. Though putting them into a single interface would
probably not be good at all.