Whee: scale quads for background blur in the shader.
This avoids the need to recalculate them every time the UI size changes
-- i.e., BaseLayer::doSetSize() would need to trigger
NeedsCompositingOffsetSizeUpdate every time either the UI size or the
framebuffer size would change, even if the actual node hierarchy
doesn't.

Now they get recalculated only when the pixel ratio changes, to adapt
the radius and smoothness expansion to that, similarly to the work
needed for regular data. The setProjection() API and shader logic is
basically verbatim taken from what's done for the BaseLayer quads
themselves.