Whee: handle the case of TextLayer text being fed into itself.
Which I assume is extremely likely to happen in any app of non-trivial
complexity. Originally I assumed it would be enough to handle this
directly in Corrade's growable array APIs, fortunately I made the text
case robust enough to uncover the need to do this also in
TextLayer::updateText() directly.