Whee: rework animation advancing internals.
While trying to integrate animation merging into the existing interface
I realized the various animator advance() implementations carry a lot of
repetitive logic (and thus also testing overhead), not to mention the
need to allocate for the delegated advance() call to have where to put
the data.

Which made me realize the animation advancing should have been done in a
way that was already invented for the layers themselves -- in
particular, splitting the "update" and "advance" steps, and have those
two driven from outside, i.e. from the AbstractUserInterface itself.

Right now, without customizable animation merging in place, the "update"
step is handled fully by the base class -- it gets the current time,
produces masks and factors based on that, and updates its internal
state in case another advance is needed again in the future. Compared to
layers & layouters, because of the internal state update all this is
being done inside AbstractAnimator instead of in
AbstractUserInterface::advanceAnimations() and then being passed in, but
I think that's fine, because unlike layers or layouters the aniamtor
doesn't rely on any outside state.

What AbstractUserInterface now does is calling the update() itself, and
then based on what it returns, calls particular animator's advance()
and/or clean(), thus the animator implementation can focus solely on the
actual animation interpolation logic, without having to worry about
correctly managing the overall state.

The remaining nastiness is in layers, where, in order to allow a layer
to perform common setup for all animators called on it, it still has to
perform all the logic for each animator on its own -- conditional
update(), conditional advance(), conditional clean() after. Not 100%
sure yet how to handle that, will finish it once I have a better idea.