Whee: include a timestamp in all events.
Will be used to drive style transition animations, and subsequently also
to implements time-dependent events such as tap and hold / long press.

As currently none of the Application classes expose a timestamp, it has
to be supplied externally. Eventually they may (Emscripten has them for
example, SDL also, although in a weird precision), so supplying them is
optional. Having the events without a timestamp might be also a way how
to opt out of an animation after a certain event, although I'm not yet
100% certain about such a use case.