Whee: allow dynamic styles to remember associated animation handle.
So when another animation / style transition is to be made on the same
data, the code can check the dynamic style ID to retrieve the previous
animation and do whatever it needs to do -- stop it, pause it, remove,
blend from it to the new animation etc.

Which quite nicely circumvents the need for general animation merging.
This is flexible enough for the styles to achieve almost anything and
doesn't require any complicated logic of maintaining or collecting lists
of animations affecting the same data and then performing a non-trivial
merging step. And, unlike the manual solution, where a widget would have
to maintain a stateful instance with all animation handles collected,
here it is associated with the layer itself, and only for the actual
animated styles, not all data instances that might maybe sometimes
animate.

The BaseLayer and TextLayer style animators now fill this information,
allowing applications to make use of this.