Whee: add UserInterfaceGL::create() and tryCreate().
Similarly to how Application classes work, because that's the
established workflow and NoCreate + setStyle() was really not intuitive
at all.

This means that for reducing confusion I also dropped the NoCreate +
size constructors, since those felt weird also ("so, is it being created
or not?"). Now the workflow is, like with Applications, either pass
everything directly to the constructor, or NoCreate and then
pass everything to create() instead.