Whee: take running style animations into account when transitioning.
This makes use of the recently added ability to associate an animation
handle with a dynamic style to perform style transitions based on input
events even for data with dynamic styles. Until now, if a particular
data had a dynamic style assigned, it was assumed that some animation or
other application-side change is made on given data and no transition
was done.

Which makes sense in general, but with the upcoming introduction of
animated style transitions, it can often happen that a style transition
is still being animated while another transition needs to be done. If
the layer would just skip data with dynamic styles, it'd mean that the
transition wouldn't be made and the visuals would be stale, for example
that a hovered & pressed data would still only look hovered because the
press was done while the hover animation was running.

So instead, if the dynamic style has an animation associated (which the
{Base,Text}LayerStyleAnimator now both do) and the animation belongs to
an animator that was set as default style animator for the layer, the
layer uses the animation target style as the source for the transition.
That then means the transition happens for those as well, ensuring the
final visual is always up-to-date. (Assuming the animation doesn't
subsequently overwrite the style index to a different one, but that's an
orthogonal problem to be solved.)