Whee: test setTransitionedStyle() called from within an event.
For press and release it behaves in a potentially surprising way where
the intermediate style is incorrect, but ultimately the final style is
correct in those two cases.

For tap or click the test uncovers a nasty corner case, where calling
this function will make given data stuck on a "pressed" style, even if
it gets released after. Fix in the next commit.