Whee: unset currentPressedNode before calling pointerTapOrClickEvent().
Fixes the issue reproduced by the test in the previous commit.
Unfortunately I didn't have an easy way to verify this inside the
AbstractUserInterface tests directly.

The currentGlobalPointerPosition suffers from the same problem, but so
far it isn't used by anything that would expose this problem, so there
it's just a TODO.