Whee: ability to create TextLayer data with a single glyph.
Intended usage is abusing the glyph cache (and thus the same draw call
as the rest of the text) for rendering also arbitrary icon and image
data.

Apart from this working also with "fonts" that are just an ID in the
glyph cache and have no actual AbstractFont instance, internally this
works the same as regular text drawing. Thus it's also possible to turn
a text TextLayer data into a single glyph data and vice versa.