Ui: put the widget style as optional last argument.
The original Ui library had this, and I didn't know the reason until I
attempted to actually use the new APIs. While it wasn't really obvious
when making the gallery app, in regular use where 90% of widgets are in
the default style, having to type it every time gets really annoying
really fast.