Ui: make it possible to NoCreate widgets.
Originally I was opposing this idea because I felt that having an
unconditional access to the UI instance is very useful, and breaking
that would make everything extremely annoying.

Fortunately, now when porting magnum-player to the new APIs, I realized
that even just NoCreate with passing the UI instance is useful enough to
have, so I did that -- and it isn't introducing any new concept either,
as this is exactly how moved-out instances look like.