Ui: UserInterfaceGL constructor and create() overloads w/ StyleFeatures.
Originally I thought this wouldn't be useful because the user only may
want to for example call setStyle() subsequently with a subset of the
style coming from elsewhere.

However, one valid use case is creating a UI with the default style but
opting out of icons for example, if the application doesn't make any use
of them, and thus also removing the need to deploy with image loading
plugins the UI would otherwise rely on.