Ui: remove {Base,Text}Layer::create*() taking color or outline width.
These were causing just too many overload combinations, and weren't
enough anyway. There's also padding and texture coordinates to be set and
those don't have any special create() overload, so why should color and
outline width have one? Ultimately it only may lead to confusion because
what's actually the extra two arguments? You can pass Vector4 / Color4 to
either, so if you pass just outline width, it'll be treated as a color,
and to do it correctly, you have to type out 0xffffff_rgbf for the color
argument, not {}, because otherwise it'll be all black.

When this gets exposed to Python, I'll use keyword-only arguments for all
the extra properties, of course. But not in C++, here it's painful, so
you get setters.