Ui: support RGBA per-data color override in {Base,Text}Layer.
Originally I thought this wouldn't be needed and I'd save four bytes per
vertex, because the assumed use case would be just custom colors for
color pickers etc., and all other cases where transparency would be
needed would be supplied through the style.

That unfortunately didn't hold up in practice, people *do* want to set
the alpha as well. Moreover, one of the planned features is an opacity
value on the nodes themselves, and in order to implement support for that
in the layers, the vertices actually need a RGBA color. So, yeah, let's
just add the alpha.