Ui: adapt to Magnum Application PointerEvent changes.
So far just the changes needed to make it work w/o deprecation warnings.
Exposing multi-touch events is coming next.

For the application event converters this means that the only causes
that originally worked for MSVC no longer do, so I'm falling back to
using the CORRADE_HAS_TYPE() helper there. For KeyEvent I'm not checking
anything, as -- so far -- it doesn't need any conflict prevention with
3rd party converters. It'll get added in a similar spirit if needed.