Ui: propagate and handle multi-touch events.
Ugh, why is anything that deals with events always such a monster
commit. Originally I thought I'd cut the multi-touch support off at the
application event translation interface and would convert those to pinch
gestures there, sending just a high-level info about two-finger zoom,
rotation or pan. But then, given how much effort went into supporting
multi-touch in the application classes already, this would be a waste of
that. So it's now being propagated all the way to the concrete data.

High-level documentation is still to be done (sigh again), but in short
it's basically the following:

 - The visual feedback on hover, press, enter/leave, etc is done only
   for primary events. Secondary presses don't have any such feedback,
   and it's up to concrete layer if or how they want to deal with them.
 - This then also implies that pointerTapOrClickEvent(),
   pointerEnterEvent() and pointerLeaveEvent() are always only getting
   primary events. Furthermore, focus-on-press is also happening for
   primary events only.
 - Secondary events also don't implicitly affect pointer capture.
   Combined with the fact that a primary (press) event makes given node
   accept remaining events until a release, this makes it possible for
   secondary events to arrive to the same data as the primary event.
 - On the other hand, secondary events still have an option to toggle
   event capture. Because why not.