Ui: use an enum set operation to check against multiple pointer types.
This is the suggested "canonical" workflow for the new Application
events, so make it here as well. The applications themselves guarantee
that the pointer() values are individual bits because they do a
translation inside, here it's guaranteed by the recently added
restrictions to PointerEvent / PointerMoveEvent constructors to enusre
that the PointerEventSource and PointerEvent values make sense together.
And because it's checked there, the rest of the code doesn't need to be
as paranoid in the checks anymore.