Ui: take node opacity into account in TextLayer.
For the actual glyphs it's the same as with BaseLayer, the per-vertex
color is multiplied with the opacity.

The cursor and selection however needed an extra attribute. Which is
extra cost, yes, but the shader is extremely simple so I didn't see any
actual difference in a benchmark, and I don't think we're bottlenecked by
the amount of selection rectangles either.