Ui: is{Hovering,Focused,Pressed}() -> isNode{Hovered,Focused,Pressed}().
I want to add a query for whether a pointer/key event is called on a
pressed node, and event.isPressed() would be way too easy to mistake for
the event press/release state. So make it explicit that those bits are
about the node, and not the event. Also the "hovering" was silly and
inconsistent, now it's "hovered" like everywhere else.