Ui: add *Event::isNodePressed().
I'm going to remove the tap/click event from AbstractUserInterface and
treat it as a "gesture" in the EventLayer, in order to be able to better
handle transitions between tap and drags, long tap, double click etc.,
which really shouldn't all be done in the main UI class.

Thus I need the extra info that the AbstractUserInterface uses to
determine whether a click really happened. Second needed bit is knowing
whether the release actually happened inside the pressed node, that's
going to be added next.

The FocusEvent and VisibilityLostEvent already had this property, adding
it to all remaining ones so it's consistent.