Ui: add {Pointer,PointerMove,Key}Event::nodeSize().
One use case is to be able to know whether the event is actually
happening inside the node or not, to decide whether e.g. a press +
release was actually a click or not, which finally allows me to move the
higher-level tap/click event from the AbstractUserInterface to
EventLayer, where they belong.

Another use case is being able to calculate distance from edges other
than top/left, for example when a visual element has rounded corners
that shouldn't react to events, or to calculate a normalized position
relative to the node size, for color pickers and such. Until now this
was only possible by explicitly remembering the actual node sizes in
doUpdate(), which means having to keep track of redundant state that's
already available elsewhere.