Ui: remove AbstractLayer::pointerTapOrClickEvent().
When doing prep design for long press, double click and especially tap
that becomes a drag I realized that the tap/click event is too high
level to be handled inside the base AbstractUserInterface class, and
should be treated as a gesture much like the pinch. Some of the reasons:

 - If a double click is implemented, the regular single tap/click should
   only be fired when it's clear that a double click won't happen. Which
   means it's insufficient / useless to get it in doPointerTapOrClick()
   and one would have to delay the call anyway.
 - If a press morphs to a drag, doPointerTapOrClick() is still called
   for a release. Again one would have to filter that out or do
   additional logic to not have both a click and a drag emitted. Doable
   (but nasty) in the context of a single event handling layer, but any
   potential other layers would still get a tap/click that they don't
   know they should ignore.
 - Similarly for a long press -- even if the pointer doesn't move,
   not turning into a drag, a long press may get fired, but a second
   tap/click gets fired again after a release.

Thus, better to leave this completely in control of the EventLayer (or,
well, any other layer) implementation. A nice side effect is the base
AbstractUserInterface getting smaller, with quite a lot less variants
and event handling corner cases to test.