Ui: make it possible to pass Application to UI creation or setSize().
While documenting the usage I realized I could just make the
UserInterface constructors take an application instance and perform the
DPI-scaling-dependent size calculation themselves. So let's do that, and
defer the actual explanation to later so it doesn't scare everyone away.

Eventually the constructors will do additional things like setting up
hooks for changing the cursor or interacting with clipboard. A nice
thing about that is that it will not require any extra work from the
user, as long as they pass the application instance it will "just work".

Existing APIs that are next to the template variants were made inline to
make it easier to verify that both are indeed doing the same thing. In
all cases they're just delegating references further, so this doesn't
really add any extra weight to the header.