Ui: implement initial opt-in event fallthrough.
It always gets out of hand, so in this case it's not yet handling
correctly the case where the currently pressed and hovered node are out
of sync. Handling that would be too nasty, so it has to wait until
pointerPressEvent() / pointerReleaseEvent() generate appropriate enter
and leave events to keep the two synchronized.

Right now it also doesn't correctly populate isNodePressed() /
isNodeHovered() where the fallthrough would happen outside of the parent
node area. This could happen for example when falling through from a
nested top-level node that's outside of its parent area, need to fix
this and add appropriate tests as well.