Ui: prevent calling {Base,Text}Layer::create(node).
The first argument is a style index, which can be any enum, which until
now could be also a NodeHandle. To a new user (or a tired user, like me
when I came across this), the following would look completely okay:

    Ui::NodeHandle node = ...;
    layer.create(node);

and it'd compile, only to likely crash at runtime because the handle bit
pattern is too large for layer style count. However, this would work and
interpret the null handle as a style with index zero:

    layer.create(Ui::NodeHandle::Null);

All create() APIs and overloads thus now explicitly check that the style
type isn't a NodeHandle.