Ui: correctly use MAGNUM_UI_BUILD_STATIC instead of MAGNUM_BUILD_STATIC.
The Ui library can be built as static while Magnum itself is still
dynamic, and the check should reflect that. OTOH, linking plugins should
happen only if they're built as static, which is likely when Magnum
itself is static.

Of course it can still happen that a subset of plugins is static and
others are dynamic, but let's not overcomplicate here.