Ui: move style out of range tests next to layer create() tests.
This is again a leftover from the times before the common style-related
code was put into AbstractVisualLayer. Now the only place in BaseLayer
and TextLayer where a style can eb out of bounds is in create(), so it
doesn't make sense to have the test so detached from those test cases.