Ui: don't override {Base,Text}Layer doFeatures() in tests for no reason.
Most probably a (copypasted) leftover from the time when there was no
base class for BaseLayer and thus the layer left implementing of
doFeatures() on the BaseLayerGL subclass.

The consequence is that the test case didn't contain a setSize() call
that's needed to be present before calling update() for any visual
layer.