Ui: switch to Source Sans OTF.
Because I want to make use of the OTF features like small caps, etc., at
the very least in the documentation but likely also in the UI itself.

Apparently, "Source Sans Pro" got renamed to "Source Sans 3", which ...
is quite amazing. Imagine how much of the world would break if Times New
Roman got renamed to Times Roman 2 in Windows 12.

Then, not sure if this is due to just the OTF and TTF handling being
different in stb_truetype, or it's actually a breaking change that
warranted the rename in the first place, the OTF output is significantly
more compact than the TTF, so basically all test files had to be
regenerated.

The OTF is also over twice as large as the TTF, but since the TTF was
added in 2017 and not updated since, it's understandable. There are
WOFF2 versions that have 150 kB instead of 320, but unfortunately, while
FreeType can open them, stb_truetype cannot, and the download size
savings I'd get from using the WOFF2 variant would be counterweighed by
switching from stb_truetype to FreeType (with Brotli for WOFF2 decoding,
etc.) in web builds. OTOH, a plain gzip compression can trim it down to
~180 kB, which isn't too different from the WOFF2, and WOFF2 isn't
really compressible further. So I guess this is still the best option.

For projects that don't make use of the builtin style / builtin widgets
I still need to add the compile-time option to exclude all this, so
long-term there's no need to worry. Finally, for very limited use it's
also possible to subset the font and include just the set of glyphs
that's actually being used in the UI. So far that's doable with
hb-subset, but I'm eyeing an option to implement this in
magnum-fontconverter directly.