Ui: try to clarify that alignment etc count has to match style count.
I also wanted to add an example snippet with a struct sliced to
individual elements, but that'd be best done in the BaseLayer instead of
duplicating in both TextLayer and once in LineLayer. However BaseLayer
doesn't yet have enough extra fields to show this on, so I'm stashing
that away for now.