Ui: accept Vector4 instead of Color4 views for LineLayer colors.
Color4 views are implicitly convertible to Vector4 views (because it's a
subclass of the same size), but the other way is not possible. So if
user code has colors as a Vector4 for some reason, it has to perform a
nasty cast here.

I also wanted to add overloads for three-component colors, but as colors
are optional, those unfortunately caused ambiguous overloads when
passing just {} or nullptr for the color view. The nullptr case could be
solved with another overload (sigh), but the {} not really. So I'm not
adding those, because the convenience benefit for three-component colors
would make the most common use case a lot more annoying. Well, I could
add an overload with no colors at all, but ... isn't that too many
overloads already?

Finally, the std::initializer_list variants stay Color4 in order to have
colors like 0x993366_rgbf implicitly converted to 0x993366ff_rgbaf,
which wouldn't be possible with just a Vector4, as that one isn't
implicitly constructible from a three-component vector for various
reasons.