Ui: use the batch glyph glyphIdsInto() API instead of a loop.
It's a lot less operations because it doesn't have to repeatedly check
the same arguments. It was done like this from the initial TextLayer
commit in d2540d1386f89d1947bf59514167c85b062f8c5f and the new
AbstractGlyphCache API was done at roughly the same time, so it's likely
that this was done before I realized a batch API would be useful, added
the batch API and then forgot to change this.