Ui: pass a glyph cache to the TextLayer::Shared constructor already.
This is how it should have been right from the start, not the extra
logic and uncertainty involved with setGlyphCache() called at some point
later or not at all.

It's a breaking change, but assuming the major use case is through the
UserInterfaceGL wrapper, which performs the TextLayer::Shared
instatiation, it shouldn't be too bad.