Ui: TextLayer glyph vertex color can be flat-interpolated.
Unlike BaseLayer or LineLayer, where the vertices can individually have
different colors to achieve various gradients, in TextLayer it's never
the case, in each glyph (and currently also the whole run) it's always
the same color for all vertices.

As the shader is dead-simple now, I don't think it has any
significant effect on performance. With the upcoming distance field
rendering it might, tho.