Ui: fix wrong RGBA color literal in test code.
This will be a compile error with an upcoming change to Magnum color
literals. And this is exactly one of the annoying bugs that are now
prevented -- it even might have hidden a bug where the texture would
shine through the semi-transparent outline, fortunately the code was
correct so it only didn't test it for regressions.