Ui: no need for these overloads.
They're needed in TextLayerGL in order to make the method chaining work
in any order with the APIs specific to TextLayerGL::Shared. But here
there are no APIs specific to BaseLayerGL::Shared so these have no use.
Same case is in LineLayerGL, it has no additional APIs so no overrides
either.