Ui: support arbitrary transformation in the TextLayer.
Originally I thought this would be done GPU-side, but I just couldn't
find a way that would be a clear win from a perf / memory use side
compared to doing this on the CPU right before upload:

 - Uploading 4 extra floats for translation + rotation to each vertex
   just to perform this operatiéon on the GPU feels extremely silly.
 - Putting this into some separate "transform" UBO like is done with
   e.g. the multi-draw-enabled Shaders is going to hit the 16 kB UBO
   limit with as little as 1024 texts. Which means there would need to
   be some extra draw call splitting logic (which is quite awful given
   that generally the draw order doesn't match the data order at all),
   or I'd have to switch to uploading just the currently visible set,
   preventing a majority of upcoming "upload only what changed"
   optimizations.
 - Anything else (querying this from a buffer texture, using a SSBO,
   ...) is either not going to improve the limit at all or improve it
   only on certain platforms, making it all only more complicated.

The next step here to improve the perf would be adding a SIMD-infused
batch transformation utility, together with potentially changing the
mesh layout to have positions all together in a dedicated stream for
better cache use when doing the transformations.

I'm planning to add a similar feature to the line rendering layer, and I
suspect it'll end up being the same, i.e. done on the CPU. Then there's
likely going to be a 3D transformation option for lines as well (such as
for rendering various gizmos), and while doing a full 3D perspective
transform is certainly more expensive than a 2D rotation + translation,
I don't think it'll be used too heavily for that to be a problem.