Ui: fill distance field TextLayer scale along with other vertex data.
It's slightly more cache friendly this way as there's a better chance
most of the vertex data are still in cache compared to doing it all
again much later, and it's less code duplication as well.

Additionally, calculate the value just once and then copy it instead of
calculating it again every time. It's likely that the compiler would
optimize this, but once it gets more involved, it might not anymore.