Ui: make AbstractUserInterface accessible from within AbstractLayer.
For (future) things like updating application cursor, accessing
clipboard and such, because it's better than doing this in an ad-hoc way
such as passing the UI reference to a custom layer constructor.

As I'm not yet sure about potential consequences of accessing and
mutating e.g. the UI node hierarchy from within critical sections such
as doUpdate(), the API is made protected, although that won't really
prevent misuse, will it.