Ui: rethink how usedCount() is calculated.
Originally I hated the way this is done because it involved jumping
around in memory, and if I had a chance to go linearly instead, I did.
But in practice the free list is always either empty or significantly
smaller than the total capacity, so it actually makes sense to iterate
just the free list and not everything.