Ui: add AbstractLayer::preUpdate().
Gets called before the actual update(), and allows layers to perform
operations that may change the set of visible nodes or data without
needing to introduce a dedicated API for those, having to go through
animators or other crazy things.

Will be used by the upcoming DebugLayer to monitor and react to changes
in the node hierarchy.