Ui: new DebugLayer class for inspecting live UIs.
So far provides "just" a way to highlight a node and see its parent /
children status as well as layer data attached to it. Later on it'll
also go through layouters and animators, but wanted to keep the initial
commit in reasonable bounds, as it's quite massive already.

No docs either, that'll happen once I have debug integration for
AbstractVisualLayer done and it's used in magnum-ui-gallery.