Ui: add DebugLayerSource::NodeAttachmentDetails.
I wanted to eventually introduce an option to print data attachments but
not their details in case it'd be for example only desirable to inspect
how much data is assigned for a particular layer and all details would
be too verbose. While it's possible to do that by passing the
AbstractLayer& base instead of concrete types, but that shifts the
burden to the user, which shouldn't be the case here, the DebugLayer
should be as convenient to use as possible.

But now I realized that adding this option later would cause breakages
to existing uses which expect just NodeAttachments to do that, so I'm
adding this differentiation right away. Besides that, it makes the
documentation easier to write, as setting a layer name does just that,
and this is another opt-in level of functionality that doesn't have to
be explained all at once. Yay.